backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
matplotlib==3.7.2
numpy==1.24.3
pymysql==1.1.0
sqlalchemy==2.0.23
pandas==2.1.4
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
faker==20.1.0
cryptography==41.0.2
langchain==0.0.354
langchain-community==0.0.20  # Downgraded to resolve langsmith conflict with langchain 0.0.354
langchain-openai==0.0.6  # Available version, not yanked
huggingface-hub==0.20.3
boto3==1.34.0
pinecone-client==3.0.0